fix: set warehouse_id/deduct stock on order creation, unify driverId to Integer, rewrite init.sql
- BUG-001: set warehouse_id from mall in createOrder
- BUG-002: deduct mall.available_quantity after order insert
- BUG-003: Long driverId → Integer driverId across Controller/Service/ServiceImpl
- BUG-004: rewrite init.sql to drop delivery_route/delivery_personnel, match current entities
- Add CLAUDE.md with bug tracking and fix records
- Add .claude/ and .npm-cache/ to .gitignore

+++ backend/src/main/java/com/logistics/controller/OrderController.java
@@ -48,7 +48,7 @@ public class OrderController {
     @Operation(summary = "查询待揽收订单")
     @GetMapping("/pending-pickup")
     public Result<List<Order>> listPendingPickupOrders(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId,
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId,
             @Parameter(description = "搜索关键词") @RequestParam(required = false) String search) {
         List<Order> orders = orderService.getPendingPickupOrders(driverId, search);
         return Result.success(orders);
@@ -82,7 +82,7 @@ public class OrderController {
     @Operation(summary = "查询待送货订单")
     @GetMapping("/pending-delivery")
     public Result<List<Order>> listPendingDeliveryOrders(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId) {
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId) {
         List<Order> orders = orderService.getPendingDeliveryOrders(driverId);
         return Result.success(orders);
     }
@@ -90,7 +90,7 @@ public class OrderController {
     @Operation(summary = "创建送货批次")
     @PostMapping("/delivery-batch")
     public Result<com.logistics.dto.CreateBatchResponse> createDeliveryBatch(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId,
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId,
             @Parameter(description = "订单ID列表") @RequestBody List<Integer> orderIds) {
         var resp = orderService.createDeliveryBatch(driverId, orderIds);
         return Result.success("创建送货批次成功", resp);
@@ -126,7 +126,7 @@ public class OrderController {
     @Operation(summary = "查询运输批次列表")
     @GetMapping("/delivery-batches-with-status")
     public Result<List<com.logistics.dto.DeliveryBatchResponse>> listDeliveryBatchesWithStatus(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId,
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId,
             @Parameter(description = "仓库ID") @RequestParam(required = false) Integer warehouseId) {
         List<com.logistics.dto.DeliveryBatchResponse> batches = orderService.getDeliveryBatchesWithStatus(driverId, warehouseId);
         return Result.success(batches);
@@ -142,7 +142,7 @@ public class OrderController {
     @Operation(summary = "查询历史任务")
     @GetMapping("/completed-batches")
     public Result<List<Order>> listCompletedBatches(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId,
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId,
             @Parameter(description = "开始时间") @RequestParam(required = false) String startTime,
             @Parameter(description = "结束时间") @RequestParam(required = false) String endTime) {
         LocalDateTime startDateTime = null;
@@ -160,7 +160,7 @@ public class OrderController {
     @Operation(summary = "查询已完成批次列表")
     @GetMapping("/completed-batches-with-status")
     public Result<List<com.logistics.dto.DeliveryBatchResponse>> listCompletedBatchesWithStatus(
-            @Parameter(description = "配送员ID") @RequestParam Long driverId,
+            @Parameter(description = "配送员ID") @RequestParam Integer driverId,
             @Parameter(description = "开始时间") @RequestParam(required = false) String startTime,
             @Parameter(description = "结束时间") @RequestParam(required = false) String endTime) {
         LocalDateTime startDateTime = null;
